+++ Example/CodeEditSourceEditorExample/CodeEditSourceEditorExample/Views/ContentView.swift
@@ -46,6 +46,7 @@ struct ContentView: View {
                 useThemeBackground: true,
                 highlightProviders: [treeSitterClient],
                 contentInsets: NSEdgeInsets(top: proxy.safeAreaInsets.top, left: 0, bottom: 28.0, right: 0),
+                additionalTextInsets: NSEdgeInsets(top: 1, left: 0, bottom: proxy.size.height * 0.3, right: 0),
                 useSystemCursor: useSystemCursor
             )
             .overlay(alignment: .bottom) {
