+++ Sources/CodeEditSourceEditor/Controller/TextViewController+FindPanelTarget.swift
@@ -10,13 +10,11 @@ import CodeEditTextView
 
 extension TextViewController: FindPanelTarget {
     func findPanelWillShow(panelHeight: CGFloat) {
-        scrollView.contentInsets.top += panelHeight
-        gutterView.frame.origin.y = -scrollView.contentInsets.top
+        updateContentInsets()
     }
 
     func findPanelWillHide(panelHeight: CGFloat) {
-        scrollView.contentInsets.top -= panelHeight
-        gutterView.frame.origin.y = -scrollView.contentInsets.top
+        updateContentInsets()
     }
 
     var emphasisManager: EmphasisManager? {
